Feature test outcome summarization for Go (#251)
Adds a tcp listener to the features CLI that accepts JSONL data
describing feature test results. The JSON objects are expected to have
at least three string fields: `name` (test name), `outcome`
(PASSED|FAILED|SKIPPED), and `message` (free text). The Go test runner
is also updated to emit this data. We use a tcp socket here to support
execution under Windows.

The overall goal here is to avoid trying to run history replays for
tests that have been skipped (e.g. due to a lack of server support for
the feature as is currently the case for older servers and the update
feature).