Allow tests to avoid the PanicPolicy override
This enum doesn't have an "unset" value so the harness cannot tell the
difference between a defaulted value and the value intentionally set to
the default. The majority of the time we want the FailWorkflow behavior
but in edge cases (e.g. tests that intentionally test panic behavior)
allow the test to opt out of the override behavior with the
DisableWorkflowPanicPolicyOverride flag.